Added two patches to let users customize the theme further with patch.
The first patch adds WAY more lines to the network monitor. This is
primarily for larger screen resolutions that have the screen real
estate, and want to see more network activity.

The second patch changes the colors of the entire Conky theme to work
well on wallpapers that are super-light. (Or white!)